My calloc

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -21,9 +21,9 @@ int _strlen(char *str)
 *Return: the concatonated stirng
 */
 
-char *string_nconcat(char *s1, char *s2, size_t n)
+char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
-	int i, j;
+	unsigned int i, j;
 	char *ptr;
 
 	if (s1 == NULL)
